+++ worker/include/Channel/ChannelNotifier.hpp
@@ -1,7 +1,6 @@
 #ifndef MS_CHANNEL_NOTIFIER_HPP
 #define MS_CHANNEL_NOTIFIER_HPP
 
-#include "common.hpp"
 #include "Channel/ChannelSocket.hpp"
 #include <string>
 
@@ -55,7 +54,7 @@ namespace Channel
 		// Passed by argument.
 		Channel::ChannelSocket* channel{ nullptr };
 		// Others.
-		flatbuffers::FlatBufferBuilder bufferBuilder{};
+		flatbuffers::FlatBufferBuilder bufferBuilder;
 	};
 } // namespace Channel