+++ worker/include/RTC/Codecs/VP8.hpp
@@ -118,7 +118,7 @@ namespace RTC
 				  : RTC::Codecs::EncodingContext(params)
 				{
 				}
-				~EncodingContext() = default;
+				~EncodingContext() override = default;
 
 				/* Pure virtual methods inherited from RTC::Codecs::EncodingContext. */
 			public:
@@ -147,7 +147,7 @@ namespace RTC
 				}
 				bool Process(
 				  RTC::Codecs::EncodingContext* encodingContext, RTC::RtpPacket* packet, bool& marker) override;
-				void RtpPacketCloned(RtpPacket* packet) override{};
+				void RtpPacketCloned(RtpPacket* packet) override {};
 				std::unique_ptr<RTC::Codecs::PayloadDescriptor::Encoder> GetEncoder() const override
 				{
 					return this->payloadDescriptor->GetEncoder();