+++ worker/src/RTC/SCTP/association/Socket.cpp
@@ -25,13 +25,15 @@ namespace RTC
 			MS_TRACE();
 		}
 
-		void Socket::Dump(int indentation) const
+		// NOLINTNEXTLINE (readability-convert-member-functions-to-static)
+		void Socket::Dump(int /*indentation*/) const
 		{
 			MS_TRACE();
 
 			// TODO
 		}
 
+		// NOLINTNEXTLINE (readability-convert-member-functions-to-static)
 		void Socket::SendInitChunk()
 		{
 			MS_TRACE();