+++ worker/src/RTC/SimpleConsumer.cpp
@@ -307,14 +307,12 @@ namespace RTC
 		// greater than computed one, then use it.
 		auto maxBitrate = this->rtpParameters.encodings[0].maxBitrate;
 
-		if (maxBitrate > desiredBitrate)
-		{
-			desiredBitrate = maxBitrate;
-		}
+		desiredBitrate = std::max(maxBitrate, desiredBitrate);
 
 		return desiredBitrate;
 	}
 
+	// NOLINTNEXTLINE (misc-no-recursion)
 	void SimpleConsumer::SendRtpPacket(RTC::RtpPacket* packet, RTC::SharedRtpPacket& sharedPacket)
 	{
 		MS_TRACE();
@@ -589,10 +587,7 @@ namespace RTC
 		auto fractionLost = this->rtpStream->GetFractionLost();
 
 		// If our fraction lost is worse than the given one, update it.
-		if (fractionLost > worstRemoteFractionLost)
-		{
-			worstRemoteFractionLost = fractionLost;
-		}
+		worstRemoteFractionLost = std::max(fractionLost, worstRemoteFractionLost);
 	}
 
 	void SimpleConsumer::ReceiveNack(RTC::RTCP::FeedbackRtpNackPacket* nackPacket)