+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ### NEXT
 
+### 3.19.9
+
+- Fix RTCP packets lost in stats ([PR #1651](https://github.com/versatica/mediasoup/pull/1651)).
+
+### 3.19.8
+
+- Fix RTCP cumulative total lost computation ([PR #1650](https://github.com/versatica/mediasoup/pull/1650)).
+
 ### 3.19.7
 
 - Bump up Meson from 1.5.0 to 1.9.1 ([PR #1634](https://github.com/versatica/mediasoup/pull/1634)).